requirements.txt
beautifulsoup4
discord.py
Flask
gspread
helper
line_bot_sdk
oauth2client
PyMySQL
Requests
